backend/main.py
"""
NERIA AI Backend Server
FastAPI server for handling queries and agent communication
"""

import os
import asyncio
import logging
from typing import Optional
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Logging setup
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(levelname)s - %(message)s"
)
logger = logging.getLogger(__name__)

# FastAPI app
app = FastAPI(
    title="NeriaMind API",
    description="Expert-Validated AI Answers",
    version="1.0.0"
)

# CORS middleware (for Next.js frontend)
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"],  # Next.js default port
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Data models
class QueryRequest(BaseModel):
    query: str
    context: Optional[str] = None

class QueryResponse(BaseModel):
    response: str
    reasoning_chain: Optional[dict] = None
    validation_status: Optional[str] = None
    capsule_id: Optional[str] = None

class KnowledgeCapsule(BaseModel):
    id: str
    query: str
    reasoning: dict
    validation: dict
    created_at: str

# Routes
@app.get("/")
async def root():
    """Health check endpoint"""
    return {
        "status": "running",
        "service": "NERIA AI",
        "version": "1.0.0"
    }

@app.post("/query", response_model=QueryResponse)
async def process_query(request: QueryRequest):
    """
    Process a reasoning query
    This will communicate with your reasoning and validation agents
    """
    logger.info(f"Received query: {request.query}")

    # TODO: Send query to reasoning agent
    # TODO: Get reasoning chain from agent
    # TODO: Send to validation agent
    # TODO: Create knowledge capsule

    # Placeholder response
    return QueryResponse(
        response="Query received and processing...",
        reasoning_chain=None,
        validation_status="pending",
        capsule_id=None
    )

@app.get("/capsules")
async def get_capsules():
    """Get all knowledge capsules"""
    # TODO: Retrieve capsules from storage
    return {"capsules": []}

@app.get("/capsules/{capsule_id}")
async def get_capsule(capsule_id: str):
    """Get a specific knowledge capsule"""
    # TODO: Retrieve specific capsule
    return {"capsule_id": capsule_id, "data": None}

if __name__ == "__main__":
    import uvicorn
    logger.info("Starting NeriaMind Backend Server...")
    uvicorn.run(app, host="0.0.0.0", port=8000)
